backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
celery>=5.3.6
django-cors-headers>=4.3.1
django-celery-beat>=2.5.0
redis>=5.0.1
pytest-django>=4.5.2
django-debug-toolbar>=4.2.0
gunicorn
psycopg2-binary
requests
